quantum-api/requirements.txt
alembic==1.4.2
antiorm==1.2.1
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
db==0.1.1
Django==3.0.2
djangorestframework==3.11.0
dnspython==1.16.0
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-cors==3.0.8
idna==2.9
img2pdf==0.3.3
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
mysqlclient==1.4.6
Pillow==7.0.0
pipenv==2018.11.26
protobuf==3.6.1
PyJWT==1.7.1
PyPDF2==1.26.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
sqlparse==0.3.0
urllib3==1.25.8
Werkzeug==1.0.1
whitenoise==5.0.1
pyquil==2.19.0